+++ .env.sample
@@ -1,3 +1,3 @@
 POSTGRES_DB=
-POSTGRES_NAME=
+POSTGRES_USER=
 POSTGRES_PASSWORD=
